pyproject.toml
[tool.poetry]
name = "tdd"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "0.85.1"
uvicorn = "0.19.0"
asyncpg = "0.27.0"
tortoise-orm = "0.19.2"
aerich = "0.7.1"
gunicorn = "20.1.0"


[tool.poetry.group.dev.dependencies]
pytest = "7.2.0"
requests = "2.28.1"
python-lsp-server = {extras = ["all"], version = "^1.7.1"}
pytest-cov = "4.0.0"
flake8 = "5.0.4"
black = "22.10.0"
isort = "5.10.1"


[tool.aerich]
tortoise_orm = "tdd.db.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
